+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Note: Can be used using nvuillam/mega-linter@insiders in your mega-linter.yml file
 
+- Fixes
+  - Fix python error when File.io does not respond, + harmonize reporter logs
+
 ## [4.14.0] 2020-12-07
 
 - Linters